code read to start fronend app

+++ src/main/java/com/communityconnect/spring/repository/UserRepository.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.communityconnect.spring.model.User;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
   Optional<User> findByEmail(String email);
-
 }
